Java university big homework
This was a big homework. Very simple, easy. We were studying the basics of programming. The window made on hungarian language. You can choose tickets for football games, and you have the opportunity to delete them, and also to order it. If you ordered, a file will contain all the informations about the order.